Revert "Fix symfony dependency"
This reverts commit d98ed7807c00f8bc53a797d6a2b3e8a0b4bfd405.

We can not change the parent class until we are SF 5.0 only and
therefore have removed all support for the legacy compatible event
dispatcher with swapped arguments.
This is needed, as the Tracable event dispatcher has a check for swapped
arguments and enforces the event class used to be a legacy event as
well.
If we change the parent class here, we do not have a legacy event class
anymore and therefore break every legacy caller.